+++ lib/src/config/config_import.dart
@@ -4,6 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:opticore/opticore.dart';
 
 import '../utils/ui/core_assets.dart';
+import '../utils/ui/core_colors.dart';
 
 part '../config/infrastructure/core_setup.dart';
 part '../config/settings/api_response_config.dart';
@@ -12,4 +13,5 @@ part '../config/settings/maintenance_config.dart';
 part '../config/settings/network_config.dart';
 part '../config/settings/no_internet_config.dart';
 part '../config/settings/not_found_config.dart';
+part '../config/settings/toast_config.dart';
 part '../config/settings/unauthenticated_config.dart';